+++ forui/example/lib/sandbox.dart
@@ -15,6 +15,7 @@ class _SandboxState extends State<Sandbox> with SingleTickerProviderStateMixin {
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   final FRadioSelectGroupController<Notification> controller = FRadioSelectGroupController();
   late FPopoverController popoverController;
+  late FCalendarController<DateTime?> a = FCalendarController.date();
 
   @override
   void initState() {
@@ -28,41 +29,7 @@ class _SandboxState extends State<Sandbox> with SingleTickerProviderStateMixin {
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            FSelectMenuTile(
-              groupController: controller,
-              popoverController: popoverController,
-              autoHide: true,
-              validator: (value) => value == null ? 'Select an item' : null,
-              prefixIcon: FIcon(FAssets.icons.bell),
-              title: const Text('Notifications'),
-              details: ListenableBuilder(
-                listenable: controller,
-                builder: (context, _) => Text(switch (controller.values.firstOrNull) {
-                  Notification.all => 'All',
-                  Notification.direct => 'Direct Messages',
-                  null || Notification.nothing => 'None',
-                }),
-              ),
-              menu: [
-                FSelectTile(title: const Text('All'), value: Notification.all),
-                FSelectTile(title: const Text('Direct Messages'), value: Notification.direct),
-                FSelectTile(title: const Text('None'), value: Notification.nothing),
-              ],
-            ),
-            const SizedBox(height: 20),
-            FButton(
-              autofocus: true,
-              label: const Text('Save'),
-              onPress: () {
-                if (!_formKey.currentState!.validate()) {
-                  // Handle errors here.
-                  return;
-                }
-
-                _formKey.currentState!.save();
-                // Do something.
-              },
-            )
+            FCalendar(controller: a),
           ],
         ),
       );