+++ forui/lib/src/widgets/calendar/calendar_controller.dart
@@ -4,9 +4,11 @@ import 'package:forui/forui.dart';
 
 bool _true(DateTime _) => true;
 
+DateTime _convert(DateTime date) => DateTime.utc(date.year, date.month, date.day);
+
 /// A controller that controls date selection in a calendar.
 ///
-/// The [DateTime]s are always in UTC timezone and truncated to the nearest day.
+/// All returned [DateTime]s are in UTC timezone with no time component.
 ///
 /// This class should be extended to customize date selection. By default, the following controllers are provided:
 /// * [FCalendarController.date] for selecting a single date.
@@ -18,64 +20,126 @@ abstract class FCalendarController<T> extends FValueNotifier<T> {
   ///
   /// [selectable] will always return true if not given.
   ///
+  /// [autoConvert] determines whether the controller should truncate and convert all given [DateTime]s to dates in
+  /// UTC timezone. Defaults to true.
+  ///
+  /// ```dart
+  /// DateTime convert(DateTime date) => DateTime.utc(date.year, date.month, date.day);
+  /// ```
+  ///
+  /// [autoConvert] should be set to false if you can guarantee that all dates are in UTC timezone (with the help of an
+  /// 3rd party library), which will improve performance. **Warning:** Giving a [DateTime] in local timezone or with a
+  /// time component when [autoConvert] is false is undefined behavior.
+  ///
   /// ## Contract
-  /// Throws [AssertionError] if [initialSelection] is not in UTC timezone.
+  /// Throws [AssertionError] if [initialSelection] is not in UTC timezone and [autoConvert] is false.
   static FCalendarController<DateTime?> date({
     DateTime? initialSelection,
     Predicate<DateTime>? selectable,
+    bool autoConvert = true,
   }) =>
-      _DateController(initialSelection: initialSelection, selectable: selectable);
+      autoConvert
+          ? _AutoDateController(initialSelection: initialSelection, selectable: selectable)
+          : _DateController(initialSelection: initialSelection, selectable: selectable);
 
   /// Creates a [FCalendarController] that allows multiple dates to be selected, with the given initial selected dates.
   ///
   /// [selectable] will always return true if not given.
   ///
+  /// [autoConvert] determines whether the controller should truncate and convert all given [DateTime]s to dates in
+  /// UTC timezone. Defaults to true.
+  ///
+  /// ```dart
+  /// DateTime convert(DateTime date) => DateTime.utc(date.year, date.month, date.day);
+  /// ```
+  ///
+  /// [autoConvert] should be set to false if you can guarantee that all dates are in UTC timezone (with the help of an
+  /// 3rd party library), which will improve performance. **Warning:** Giving a [DateTime] in local timezone or with a
+  /// time component when [autoConvert] is false is undefined behavior.
+  ///
   /// ## Contract
-  /// Throws [AssertionError] if the dates in [initialSelections] are not in UTC timezone.
+  /// Throws [AssertionError] if the dates in [initialSelections] are not in UTC timezone and [autoConvert] is false.
   static FCalendarController<Set<DateTime>> dates({
     Set<DateTime> initialSelections = const {},
     Predicate<DateTime>? selectable,
+    bool autoConvert = true,
   }) =>
-      _DatesController(initialSelections: initialSelections, selectable: selectable);
+      autoConvert
+          ? _AutoDatesController(initialSelections: initialSelections, selectable: selectable)
+          : _DatesController(initialSelections: initialSelections, selectable: selectable);
 
   /// Creates a [FCalendarController] that allows a single range to be selected, with the given initial range.
   ///
   /// [selectable] will always return true if not given.
   ///
-  /// Both the start and end dates of the range is inclusive. The selected dates are always in UTC timezone and
-  /// truncated to the nearest day. Unselectable dates within the selected range are selected regardless.
+  /// [autoConvert] determines whether the controller should truncate and convert all given [DateTime]s to dates in
+  /// UTC timezone. Defaults to true.
+  ///
+  /// ```dart
+  /// DateTime convert(DateTime date) => DateTime.utc(date.year, date.month, date.day);
+  /// ```
+  ///
+  /// [autoConvert] should be set to false if you can guarantee that all dates are in UTC timezone (with the help of an
+  /// 3rd party library), which will improve performance. **Warning:** Giving a [DateTime] in local timezone or with a
+  /// time component when [autoConvert] is false is undefined behavior.
+  ///
+  /// Both the start and end dates of the range is inclusive. Unselectable dates within the selected range are
+  /// selected regardless.
   ///
   /// ## Contract
   /// Throws [AssertionError] if:
-  /// * the given dates in [value] is not in UTC timezone.
+  /// * the given dates in [value] is not in UTC timezone and [autoConvert] is false.
   /// * the end date is less than start date.
   static FCalendarController<(DateTime, DateTime)?> range({
     (DateTime, DateTime)? initialSelection,
     Predicate<DateTime>? selectable,
+    bool autoConvert = true,
   }) =>
-      _RangeController(initialSelection: initialSelection, selectable: selectable);
+      autoConvert
+          ? _AutoRangeController(initialSelection: initialSelection, selectable: selectable)
+          : _RangeController(initialSelection: initialSelection, selectable: selectable);
 
   /// Creates a [FCalendarController] with the given initial [value].
   FCalendarController(super._value);
 
   /// Returns true if the given [date] can be selected/unselected.
   ///
-  /// [date] should always in UTC timezone and truncated to the nearest day.
-  ///
   /// ## Note
   /// It is unsafe for this function to have side effects since it may be called more than once for a single date. As it
-  /// is called frequently, it should not be computationally expensive.
+  /// is called frequently, it should also not be computationally expensive.
   bool selectable(DateTime date);
 
   /// Returns true if the given [date] is selected.
-  ///
-  /// [date] should always in UTC timezone and truncated to the nearest day.
   bool selected(DateTime date);
 
   /// Selects the given [date].
-  ///
-  /// [date] should always in UTC timezone and truncated to the nearest day.
   void select(DateTime date);
+}
+
+// The single date controllers.
+class _AutoDateController extends FCalendarController<DateTime?> {
+  final Predicate<DateTime> _selectable;
+
+  _AutoDateController({
+    DateTime? initialSelection,
+    Predicate<DateTime>? selectable,
+  })  : _selectable = selectable ?? _true,
+        super(initialSelection = initialSelection == null ? null : _convert(initialSelection));
+
+  @override
+  bool selectable(DateTime date) => _selectable(_convert(date));
+
+  @override
+  bool selected(DateTime date) => value == _convert(date);
+
+  @override
+  void select(DateTime date) {
+    date = _convert(date);
+    super.value = value == date ? null : date;
+  }
+
+  @override
+  set value(DateTime? value) => super.value = value == null ? null : _convert(value);
 }
 
 class _DateController extends FCalendarController<DateTime?> {
@@ -92,12 +156,36 @@ class _DateController extends FCalendarController<DateTime?> {
   bool selectable(DateTime date) => _selectable(date);
 
   @override
-  bool selected(DateTime date) => value?.toLocalDate() == date.toLocalDate();
+  bool selected(DateTime date) => value == date;
+
+  @override
+  void select(DateTime date) => value = value == date ? null : date;
+}
+
+// The multiple dates controllers.
+final class _AutoDatesController extends FCalendarController<Set<DateTime>> {
+  final Predicate<DateTime> _selectable;
+
+  _AutoDatesController({
+    Set<DateTime> initialSelections = const {},
+    Predicate<DateTime>? selectable,
+  })  : _selectable = selectable ?? _true,
+        super(initialSelections.map(_convert).toSet());
+
+  @override
+  bool selectable(DateTime date) => _selectable(_convert(date));
+
+  @override
+  bool selected(DateTime date) => value.contains(_convert(date));
 
   @override
   void select(DateTime date) {
-    value = value?.toLocalDate() == date.toLocalDate() ? null : date;
+    final copy = {...value};
+    super.value = copy..toggle(_convert(date));
   }
+
+  @override
+  set value(Set<DateTime> value) => super.value = value.map(_convert).toSet();
 }
 
 final class _DatesController extends FCalendarController<Set<DateTime>> {
@@ -117,10 +205,66 @@ final class _DatesController extends FCalendarController<Set<DateTime>> {
   bool selected(DateTime date) => value.contains(date);
 
   @override
-  void select(DateTime date) {
-    final copy = {...value};
-    value = copy..toggle(date);
+  void select(DateTime date) => value = value..toggle(date);
+}
+
+// The range controllers.
+final class _AutoRangeController extends FCalendarController<(DateTime, DateTime)?> {
+  final Predicate<DateTime> _selectable;
+
+  _AutoRangeController({
+    (DateTime, DateTime)? initialSelection,
+    Predicate<DateTime>? selectable,
+  })  : _selectable = selectable ?? _true,
+        super(
+          initialSelection =
+              initialSelection == null ? null : (_convert(initialSelection.$1), _convert(initialSelection.$2)),
+        ) {
+    final range = value;
+    assert(
+      range == null || (range.$1.isBefore(range.$2) || range.$1.isAtSameMomentAs(range.$2)),
+      'end date must be greater than or equal to start date',
+    );
   }
+
+  @override
+  bool selectable(DateTime date) => _selectable(_convert(date));
+
+  @override
+  bool selected(DateTime date) {
+    if (value case (final first, final last)) {
+      final current = date.toLocalDate();
+      return first.toLocalDate() <= current && current <= last.toLocalDate();
+    }
+
+    return false;
+  }
+
+  @override
+  void select(DateTime date) {
+    date = _convert(date);
+    if (value == null) {
+      super.value = (date, date);
+      return;
+    }
+
+    final (first, last) = value!;
+
+    switch ((first, last)) {
+      case (final first, final last) when date == first || date == last:
+        super.value = null;
+
+      case (final first, final last) when date.isBefore(first):
+        super.value = (date, last);
+
+      case (final first, _):
+        super.value = (first, date);
+    }
+  }
+
+  @override
+  set value((DateTime, DateTime)? value) =>
+      super.value = value == null ? null : (_convert(value.$1), _convert(value.$2));
 }
 
 final class _RangeController extends FCalendarController<(DateTime, DateTime)?> {
@@ -158,22 +302,21 @@ final class _RangeController extends FCalendarController<(DateTime, DateTime)?> 
   @override
   void select(DateTime date) {
     if (value == null) {
-      value = (date, date);
+      super.value = (date, date);
       return;
     }
 
     final (first, last) = value!;
-    final pressed = date.toLocalDate();
 
-    switch ((first.toLocalDate(), last.toLocalDate())) {
-      case (final first, final last) when pressed == first || pressed == last:
-        value = null;
+    switch ((first, last)) {
+      case (final first, final last) when date == first || date == last:
+        super.value = null;
 
-      case (final first, final last) when pressed < first:
-        value = (pressed.toNative(), last.toNative());
+      case (final first, final last) when date.isBefore(first):
+        super.value = (date, last);
 
       case (final first, _):
-        value = (first.toNative(), pressed.toNative());
+        super.value = (first, date);
     }
   }
 }